backend/ia-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
gunicorn==21.2.0
requests==2.31.0
urllib3==2.0.4
